+++ src/main/java/be/baur/sds/content/DecimalType.java
@@ -1,6 +1,6 @@
 package be.baur.sds.content;
 
-import be.baur.sds.common.Content;
+import be.baur.sds.serialization.Content;
 
 /**
  * A <code>DecimalType</code> represents an SDA node with decimal content.<br>