+++ src/main/java/be/baur/sds/content/IntegerType.java
@@ -1,6 +1,6 @@
 package be.baur.sds.content;
 
-import be.baur.sds.common.Content;
+import be.baur.sds.serialization.Content;
 
 /**
  * A <code>IntegerType</code> represents an SDA node with integer content.<br>